Change name of the classes & Change name of the components
