+++ packages/frontend/src/components/navbar/Navbar.js
@@ -13,7 +13,7 @@ import {
 
 import './Navbar.scss'
 import './NavbarMobileMenu.scss'
-import Network from './Network'
+import NetworkSelect from './NetworkSelect'
 
 const links = [
   { title: 'Home', href: '/' },
@@ -41,62 +41,62 @@ const NavLink = (props) => {
       }}
       color={'white'}
       href={to}
-      whiteSpace={'nowrap'}    
-        >
+      whiteSpace={'nowrap'}
+    >
       {children}
     </Box>
   )
 }
 
-function Navbar () {
+function Navbar() {
   const { isOpen, onOpen, onClose } = useDisclosure()
 
   return (
     <Box px={3}>
-        <Flex
-          className={'Navbar'}
-          maxW='1980px'
-          ml='auto'
-          mr='auto'
-          h={16}
-          alignItems={'center'}
-          justifyContent={'space-between'}
-        >
-          <IconButton
-            size={'md'}
-            icon={isOpen ? <CloseIcon /> : <HamburgerIcon />}
-            aria-label={'Open Menu'}
-            display={{ lg: 'none' }}
-            marginRight={'8px'}
-            onClick={isOpen ? onClose : onOpen}
-          />
+      <Flex
+        className={'Navbar'}
+        maxW='1980px'
+        ml='auto'
+        mr='auto'
+        h={16}
+        gap='8px'
+        alignItems={'center'}
+        justifyContent={'space-between'}
+      >
+        <IconButton
+          size={'md'}
+          icon={isOpen ? <CloseIcon /> : <HamburgerIcon />}
+          aria-label={'Open Menu'}
+          display={{ lg: 'none' }}
+          onClick={isOpen ? onClose : onOpen}
+        />
 
-            <HStack spacing={8} alignItems={'center'}>
-                <HStack as={'nav'} spacing={3} display={{ base: 'none', lg: 'flex' }}>
-                {links.map((link) => (
-                    <NavLink to={link.href} key={link.title}>{link.title}</NavLink>
-                ))}
-                </HStack>
+        <HStack spacing={8} alignItems={'center'}>
+          <HStack as={'nav'} spacing={3} display={{ base: 'none', lg: 'flex' }}>
+            {links.map((link) => (
+              <NavLink to={link.href} key={link.title}>{link.title}</NavLink>
+            ))}
+          </HStack>
 
-            </HStack>
-            <div className='wrapperNetworkAndSearch'>
-              <Network />
-            <Box ml={2}>
-              <GlobalSearchInput />
-            </Box>
-            </div>
-           
-        </Flex>
+        </HStack>
+        <div className='Navbar__WrapperNetworkSelect'>
+          <NetworkSelect />
+          <Box ml={2}>
+            <GlobalSearchInput />
+          </Box>
+        </div>
 
-        {isOpen
-          ? <Box className={'NavbarMobileMenu'} pb={4} display={{ lg: 'none' }}>
-              <Stack as={'nav'} spacing={4}>
-                {links.map((link) => (
-                  <NavLink to={link.href} key={link.title}>{link.title}</NavLink>
-                ))}
-              </Stack>
-            </Box>
-          : null}
+      </Flex>
+
+      {isOpen
+        ? <Box className={'NavbarMobileMenu'} pb={4} display={{ lg: 'none' }}>
+          <Stack as={'nav'} spacing={4}>
+            {links.map((link) => (
+              <NavLink to={link.href} key={link.title}>{link.title}</NavLink>
+            ))}
+          </Stack>
+        </Box>
+        : null}
     </Box>
   )
 }
